feat: What Am I doing with this github repo

+++ src/main/java/hse/bank/records/BankAccountData.java
@@ -1,4 +1,4 @@
 package hse.bank.records;
 
-public record BankAccountData(String name, int initialBalance) implements ObjectData {
+public record BankAccountData(String name, double initialBalance) implements ObjectData {
 }
