+++ src/main/java/hse/bank/records/OperationData.java
@@ -4,6 +4,6 @@ import hse.bank.domains.BankAccount;
 import hse.bank.domains.Category;
 import java.util.Optional;
 
-public record OperationData(boolean type, BankAccount account, int amount, Category category, Optional<String> desc) implements ObjectData {
+public record OperationData(boolean type, BankAccount account, double amount, Category category, Optional<String> desc) implements ObjectData {
 
 }
